fix jsonfield bug! pip install "django-jsonfield" instead of "jsonfield". And now my Mass_Clothes works correctly

+++ django/WashComing/st.sh
@@ -1,3 +1,12 @@
-rm nohup.out 1>&2
-nohup python manage.py runserver 0.0.0.0:8002 > dj.log &
-echo $! > dj.pid
+if [ -f dj.pid ];then
+    ps -p $(cat dj.pid)
+    if [ $? -ne 0 ];then
+        rm nohup.out 1>&2
+        nohup python manage.py runserver 0.0.0.0:8002 > dj.log &
+        echo $! > dj.pid
+    fi
+else
+    rm nohup.out 1>&2
+    nohup python manage.py runserver 0.0.0.0:8002 > dj.log &
+    echo $! > dj.pid
+fi
